pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentguard"
version = "0.1.0"
description = "EU AI Act compliance middleware for AI agents. Make any LLM-powered agent legally deployable in Europe with 3 lines of code."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Sagar", email = "" },
]
keywords = [
    "eu-ai-act",
    "ai-compliance",
    "ai-agents",
    "transparency",
    "audit",
    "governance",
    "llm",
    "middleware",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
langchain = ["langchain-core>=0.2", "langchain-openai>=0.1"]
dotenv = ["python-dotenv>=1.0"]
all = ["openai>=1.0", "anthropic>=0.30", "langchain-core>=0.2", "langchain-openai>=0.1", "python-dotenv>=1.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.urls]
Homepage = "https://github.com/sagar/agentguard"
Documentation = "https://agentguard.dev"
Repository = "https://github.com/sagar/agentguard"
Issues = "https://github.com/sagar/agentguard/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
